replace free space with custom command

+++ src/main/kotlin/net/ickis/deluge/request/FreeSpace.kt
@@ -1,5 +1,5 @@
 package net.ickis.deluge.request
 
-object FreeSpace : Request<Int>("core.get_free_space") {
+object FreeSpace : Request<Long>("core.get_free_space") {
     override val args: List<Any> = emptyList()
 }
